requirements.txt
Django>=5.2.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=23.5
Pillow>=10.0.0
python-decouple>=3.8
